What:   * reduce code Why:   * 代码结构乱   * 某些函数似乎根本不应该放在那 How:   * 移

+++ inc/pvz_offset.h
@@ -42,7 +42,7 @@ struct pvz_offset pvz_off_tbl[] = {
 off_t getOffset(const char *name) {
   off_t off = -1;
   struct pvz_offset *pf;
-  for (unsigned i = 0; i < sizeof(pvz_off_tbl) / sizeof(pvz_off_tbl[0]); ++i) {
+  for (size_t i = 0; i < sizeof(pvz_off_tbl) / sizeof(pvz_off_tbl[0]); ++i) {
     pf = &pvz_off_tbl[i];
     if (strcmp(name, pf->name) == 0) {
       off = pf->offset;
